+++ package.json
@@ -15,7 +15,8 @@
   },
   "dependencies": {
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.5.1"
   },
   "homepage": "https://Daeun-100.github.io/react-payments",
   "devDependencies": {
