+++ src/pages/CardRegistrationComplete.tsx
@@ -1,0 +1,9 @@
+const CardRegistrationComplete = () => {
+  return (
+    <>
+      <div>등록 완료!</div>
+    </>
+  );
+};
+
+export default CardRegistrationComplete;
